Patch

+++ Controllers/BlogController.php
@@ -19,7 +19,7 @@ class BlogController extends Controller
     public function index()
     {
         $postModel = new PostsModel();
-        $posts     = $postModel->getAllPostWithAuthorName();
+        $posts = $postModel->getAllPostWithAuthorName();
 
         // TODO: Récupérer chaque auteur pour chaque post
 
@@ -34,7 +34,7 @@ class BlogController extends Controller
     public function post($id)
     {
         $postModel = new PostsModel();
-        $post      = $postModel->getPostWithAuthorName($id);
+        $post = $postModel->getPostWithAuthorName($id);
         $postModel->hydrate($post);
 
 
@@ -42,7 +42,7 @@ class BlogController extends Controller
         $postModel->setContent($Parsedown->text($postModel->getContent()));
 
         $commentsModel = new CommentsModel();
-        $comments      = $commentsModel->getCommentWithAuthorName($id);
+        $comments = $commentsModel->getCommentWithAuthorName($id);
 
         $this->twigRender(
             'frontoffice/postView.twig',
@@ -57,29 +57,33 @@ class BlogController extends Controller
      */
     public function editPost($id)
     {
-        // FIXME: Utilisable uniquement si administrateur
-        $postModel = new PostsModel();
-        $post      = $postModel->find($id);
+        if ($this->checkPathPrivilege('admin')) {
+            $postModel = new PostsModel();
+            $post = $postModel->find($id);
 
-        $this->twigRender(
-            'backoffice/editPostView.twig',
-            array('post' => $post)
-        );
+            $this->twigRender(
+                'backoffice/editPostView.twig',
+                array('post' => $post)
+            );
+        } else {
+            header('Location: /');
+        }
     }
 
     public function updatePost()
     {
-        // Récuperer les informations du post
-        if (isset($_POST['post'])) {
-            $data                     = $_POST['post'];
+        if ($this->checkPathPrivilege('admin') && isset($_POST['post'])) {
+            $data = $_POST['post'];
             $data['modificationDate'] = date('Y-m-d H:i:s');
 
             $postModel = new PostsModel();
-            $post      = $postModel->hydrate($data);
+            $post = $postModel->hydrate($data);
 
             $post->update();
 
             header('Location: /blog/post/' . $post->getId());
+        } else {
+            header('Location: /');
         }
     }
 
@@ -88,29 +92,33 @@ class BlogController extends Controller
         if ($this->checkPathPrivilege('admin')) {
             // Récuperer les informations du post
             if (isset($_POST['post'])) {
-                $data                    = $_POST['post'];
+                $data = $_POST['post'];
                 $data['publicationDate'] = date('Y-m-d H:i:s');
-                $data['authorId']        = $_SESSION['user']['id'];
+                $data['authorId'] = $_SESSION['user']['id'];
 
                 $postModel = new PostsModel();
-                $post      = $postModel->hydrate($data);
+                $post = $postModel->hydrate($data);
 
                 $post->create();
 
                 header('Location: /blog');
             }
         } else {
-            echo 'Erreur';
+            header('Location: /');
         }
     }
 
     public function deletePost($id)
     {
-        $postModel = new PostsModel();
-        $postModel->delete($id);
+        if ($this->checkPathPrivilege('admin')) {
+            $postModel = new PostsModel();
+            $postModel->delete($id);
 
 
-        header('Location: /users/adminDashboard');
+            header('Location: /users/adminDashboard');
+        } else {
+            header('Location: /');
+        }
     }
 
     public function addComment($post_id)
@@ -118,37 +126,37 @@ class BlogController extends Controller
         if ($this->checkPathPrivilege('admin')) {
             if (isset($_POST['comment'])) {
                 $data = array(
-                    'authorId'    => $_SESSION['user']['id'],
-                    'postId'      => (int)$post_id,
-                    'isApproved'  => 1,
-                    'comment'     => $_POST['comment'],
+                    'authorId' => $_SESSION['user']['id'],
+                    'postId' => (int)$post_id,
+                    'isApproved' => 1,
+                    'comment' => $_POST['comment'],
                     'commentDate' => date('Y-m-d H:i:s'),
                 );
 
 
                 $commentModel = new CommentsModel();
-                $comment      = $commentModel->hydrate($data);
+                $comment = $commentModel->hydrate($data);
 
                 $comment->create();
             }
         } elseif ($this->checkPathPrivilege('user')) {
             if (isset($_POST['comment'])) {
                 $data = array(
-                    'authorId'    => $_SESSION['user']['id'],
-                    'postId'      => (int)$post_id,
-                    'isApproved'  => 0,
-                    'comment'     => $_POST['comment'],
+                    'authorId' => $_SESSION['user']['id'],
+                    'postId' => (int)$post_id,
+                    'isApproved' => 0,
+                    'comment' => $_POST['comment'],
                     'commentDate' => date('Y-m-d H:i:s'),
                 );
 
 
                 $commentModel = new CommentsModel();
-                $comment      = $commentModel->hydrate($data);
+                $comment = $commentModel->hydrate($data);
 
                 $comment->create();
             }
         } else {
-            echo 'Erreur';
+            header('Location: /');
         }
         header('Location: /blog/post/' . $post_id);
     }
